fix-1819, documentation only
# Analysis #

Issue #1819 described a sentence referring to "the HTML <img /> tag" (_packages/mjml-image/README.md_). Changed to "the HTML <img> tag."

For the instances of code in our documentation that runs in modern browsers (many calls to TRY-IT-LIVE, for example), we might consider removing the deprecated closing syntax in advance of the day it might become mandatory. Those changes are here.

* _No change to any JavaScript file._

# Similar Previous Commit: PR#1567 #

PR#1567 contained the similar content as here, but applied itself to `mjmlio:master`. I closed PR#1567.