potential changes to file "doc/guide.md", documentation only
Perhaps we'll prefer our "guide.md" file with some minor edits.

# Included #

A revised "doc/guide.md file". Perhaps all tweaks and additions
are minor. Some result from feedback provided by the well-known
and valuable (and free) English grammar helper program "Grammarly".

Context: This file is the first in the series of files that are
our documentation. It's a perfect place to emphasize that MJML
is a Mailjet contribution to the industry. I like the way the
words do that.

# Self-Closing Tags #

In a November Slack conversation, Nico and Garry considered whether
the documentation should include the statement about self-closing
opening tags in "Basic layout example" > "2 Columns Section".
Probably both of us could easily feel this is common enough knowledge
among our users that it contributes little to our documentation.
I told him I'd try something; this fulfills a commitment.

Maybe this is a better place to put the information than the existing.
Maybe it'll be useful that we test the idea here.