Fixed the unclickable area in MSO-PC for the buttons
Removed the mso-padding-alt in getStyles.td, as inner padding cause is no longer needed when the size of the button is controlled by the padding in the anchor.
Removed the mso-padding-alt: 0px in getStyles.content, cause is nullifying the effect of activating the anchor as a button in MSO-PC.
Added border 1px solid which activates a "bug" in MSO-PC, forcing MSO to read the display attribute of the button (default: inline-block).
Added the this.getAttribute('background-color') property to the style, to avoid non-corresponding colours between the background of the container TD and the border created in the anchor.