Export jsonToXML helper to be used by components
Component `render()` function must return a valid XML string.
To achieve that, programmers can manipulate xml strings, it's humain
readable but hard to test.
To be more programmable and testable, component creators should be able to use
mjml json notation and then call the `jsonToXML` just before returning
the component with the `render()` method.

I also improved a bit the `jsonToXML` helper by allowing the
`attributes` field to be missing.
It allows the generated JSON to be a lot shorter.